added the ROI primary metric
Added the "ROI" metric - return on investment - as a new column in the loans_to_2008 dataframe. Because I didn't think the "Successful" / "Unsuccessful" metric (based solely on whether or not a loan had been completed) contained enough information for a very meaningful analysis. I also discussed the results of my univariate graphs.