spearman analyses, credit score range graphs
I added numeric representations of some of the categorical variables, and use these new columns to run spearman analyses. 
The results were largely disappointing. 
Neither of my metrics (ROI (a return on investment metric I calculated), nor NewLoanStatus (also a variable of my creation -- it assesses whether or not a loan is successful based on whether or not it was completed) matched very well with any of my predictor variables. 
I also fixed my credit score ranges versus credit grade graphs. (I'd mistakenly converted one of the credit score range columns into a categorical variable; I undid this error). However, there is a new problem with this pair of graphs. For some reason, the second of the two graphs is very small.